requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
Babel==2.9.1
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
frozenlist==1.3.1
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.2
outcome==1.2.0
pycodestyle==2.9.1
pycparser==2.21
PyMySQL==1.0.2
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wsproto==1.2.0
yarl==1.8.1
